mac80211: add cookie to remain_on_channel callback
a roc might get cancelled (by the user) after it was
expired (by the low-level driver), but before mac80211
processed the hw_roc_done work.

In this case, the hw_roc_done work might expire the
roc that has just started (when the previous roc
was cancelled), which will get mac80211 and the
low-level driver to go out of sync.

Fix it by passing the pointer of the roc struct
as cookie, and check this cookie on expiry
notification.

TODO: add flush_work, fix traces, ...
a better solution should be implemented completely
in mac80211, but the locking semantics complicate
it, so leave it for now.

Signed-off-by: Eliad Peller <eliad@wizery.com>