wlcore: generate version.h from original git in case of backports (INTERNAL)
if we use backports to compile our driver, the compilation
is done on the target kernel, but we want to generate the
version tag from the driver git.

check for the existence of KLIB_BUILD and GIT_TREE params,
and generate version.h from the GIT_TREE if they both
exist.
(note that backports doesn't require GIT_TREE to be
defined, so it has to be part of the build script)

Signed-off-by: Eliad Peller <eliad@wizery.com>