mesh: Fix WARNING in wl18xx_sta_rc_update()
Enable to use the 'sta_rc_update' opcode for mesh IF as well as for STA IF.

Also, add optimisation to update the FW with the HT mode change
or disconnect in oder to update the HT mode, only if there is a real change.
This will prevent from triggering disconnection when the 'sta_rc_update' opcode
was triggered for other rate change and not the HT mode change,
which is the only change that our driver handles.

fixes #WILINK8_SW_DEV-411

Signed-off-by: Maital Hahn <maitalm@ti.com>
Acked-by: Yaniv Machani <yanivma@ti.com>