mesh: Mesh Point always asserts "HT Mixed Mode" protection [WORK-AROUND]
There are 3 diffrences between an AP beacon and a mesh beacon, all of
them are in the HT capabilities and HT information IEs.
These 2 IEs are also diffrent between the mesh beacon and the mesh action frames.
These diffrences results from the fact that the AP beacon and the mesh action frames
are both being set/built by the hostap package, while the mesh beacon is being built
by the mac80211 layer in the kernel.

In order to align the mesh beacon to the action frames (and to the AP beacon),
we changed the default parameters in the kernel to metch the hostap defaults.

(The correct fix will be to set these IEs from the hostap also for
the mesh beacons, as it is done for the AP beacon).

fixes #WILINK8_SW_DEV-69

Signed-off-by: Maital Hahn <maitalm@ti.com>
Acked-by: Yaniv Machani <yanivma@ti.com>