mesh: Add driver capability to calculate airtime link metric correctly.
Driver will now maintain a table which
specifies a rate provided by the FW.

Also, the mac80211 layer now has the ability to
ask the wlcore for specific rate per link from
the table.

This rate is obtained by the FW during TX/RX,
and will be passed to the driver to allow it to calculate
the best estimation for airtime link metric.

Signed-off-by: Maxim Altshul <maxim.altshul@ti.com>