mesh : PREQ: increase the PREQ lifetime to imporve performance
All MPs in the mesh, except the initiator MP, after reception of PREQ or PREP,
will update a relevant entry in the mesh path table,
and set the new entry's counter to the lifetime advertised in the PREQ/PREP plus some extra time

This extra time will avoid the intermediate MPs' path expiry prior to the expiry at the path initiator MP;
This extra time will avoid the target MP path expiry,
prior to the expiry at the path initiator MP.
Ttarget MP will not start its own new path selection,
which will lead to an instant path switching, of the "main" data initiator, to a shortest path;
The extra time is set to a twice of "path_refresh_time" that by default is 1 second

The mesh path entry's lifetime will not be increased only if MP is an originator of
the path selection and if the table's entry is equal to the destination address.
For all other entries of the originator's MP the lifetime will be increase, as for
non initiator MPs.

Signed-off-by: Victor Asovsky <victora@ti.com>