mesh: update path table only with better PREQ path to next hop
When the TA of the PREQ already appears in the table as the destination,
but the current next hop to it is a different node we should compare
the metric before we replace the next hop for this destination in the table.
If this PREQs' metric is worst, we will ignore it despite the newer sequence number that it conatains.

Check that the "next hop" is not NULL
Mesh path table's entry, including its next hop update will be updated on a regular
basis in case when originator's SN advances by more then one.
Improves cases of fast healing and very congested air at the intermediate MP.

This will prevent a case of momentarily switching to the worst path.

Signed-off-by: Victor Asovsky <victora@ti.com>