mesh: TO is always enabled. Only targets responmd to Originator
TO Flag will always be set to '1'.
Prevent MPs in the middle to send PREP to the originator
in case that the originator re-sends PREQ with a newer orig_sn,
but the same target_sn and those MPs have an updated info.
Only a destination will respond to originator's request.
Prevents sequence numbers' rase between an intermediate MPs

Signed-off-by: Victor Asovsky <victora@ti.com>