mesh: Don't increase lifetime only for the current initiator's process target [SQUASH]
Squash for mesh : PREQ: increase the PREQ lifetime to imporve performance
a352752dea2ccc8c755d0cbbd940fe2ed3279caf

The mesh path entry's lifetime will not be increased only if MP is an originator of
the path selection and if the table's entry is equal to the destination address.
For all other entries of the originator's MP the lifetime will be increase, as for
non initiator MPs.

Signed-off-by: Victor Asovsky <victora@ti.com>