mesh: increase PREQ/PRES action frames reliability
Although PRES action frame is sent as unicast, it might not arrive to
it's destination.
To reduce the probability of losing this frame this fix:
- Increase the number of retries for management frames.
- Set 6Mbps rate as basic rate for management frames.

PREQ is sent as broadcast, and might collide in the air.
To reduce the probability of missing that frame, sent it twice.
SN and PDID will not be changed for PREQs of the same transmition, i.e.,
two bcast PREQs a single path selection instance will be identical

The frame will be sent twice only at the originator side,
once the second frame is received in one of the peers, it will be ignore
as the sequence number remains the same.
(This prevents a broadcast storms in the generic case)

Signed-off-by: Victor Asovsky <victora@ti.com>
Signed-off-by: Yaniv Machani <yanivma@ti.com>